Fix stack corruption in the <X> extended accessory command
RegisterList::setExtendedAccessory(char *) declared the value as a byte and
read it with sscanf("%d"), which writes sizeof(int) bytes. Every <X ADDRESS
VALUE> command therefore clobbered one byte of the stack next to it.

Reported by the compiler as:
  warning: format '%d' expects argument of type 'int*',
           but argument 4 has type 'byte* {aka unsigned char*}' [-Wformat=]

The value is passed to setExtendedAccessory(int, int) and truncated to a byte
when it is placed in the packet, so widening the local changes nothing else.

+++ PacketRegister.cpp
@@ -290,7 +290,8 @@ void RegisterList::setExtendedAccessory(int aAdd, int val) volatile
 void RegisterList::setExtendedAccessory(char *s) volatile
 {
   int aAdd; // the accessory address (1-2044 = 11 bits)
-  byte val; // the accessory value for that address (0-31) following NMRA recommended convention, but byte holds 0..255
+  int val;  // the accessory value for that address (0-31) following NMRA recommended convention, truncated to a byte in the packet
+            // NOTE: must be an int, sscanf("%d") writes sizeof(int) bytes
 
   if (sscanf(s, "%d %d", &aAdd, &val) != 2)
   {
